--- gitcd/controller/untitled
@@ -1,1 +1,0 @@
-extension=gearman.so
